Implemented small cluster and redundancy example.
This ansible playbook contains a methodology to define clusters
in Ansible inventories.

Basically it's a tree based on a given group. This inventory
contains webservers located in 3 clusters (Paris, Brussels, Madrid).
All of these webservers should be configured quiet the same excepted
for the some parameters linked to the cluster.